git-svn-id: svn://svn.chromium.org/chrome/branches/background_tasks/src@3919 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/glue/npruntime_util.cc
@@ -10,7 +10,7 @@
 #if USE(V8_BINDING)
 #include "webkit/port/bindings/v8/np_v8object.h"
 #elif USE(JAVASCRIPTCORE_BINDINGS)
-#include "bindings/c/c_utility.h"
+#include "bridge/c/c_utility.h"
 using KJS::Bindings::PrivateIdentifier;
 #endif
 
@@ -52,7 +52,7 @@ bool DeserializeNPIdentifier(const Pickle& pickle, void** pickle_iter,
     int data_len;
     if (!pickle.ReadData(pickle_iter, &data, &data_len))
       return false;
-    DCHECK_EQ(data_len, strlen(data) + 1);
+    DCHECK_EQ((static_cast<size_t>(data_len)), strlen(data) + 1);
     *identifier = NPN_GetStringIdentifier(data);
   } else {
     int number;
